fix: Removes content that is being reworked anyway from producer shelf, creates a drag wrapping component for the shelf
