+++ src/app/draggable-shelf/draggable-shelf.component.html
@@ -1,0 +1,6 @@
+<div [style.flex-basis.px]="dragOffsetInPixels" [style.height.px]="dragOffsetInPixels">
+    <div #dragHandle class="drag-handle">
+        <sofie-icon-button [iconButton]="IconButton.BARS" [iconButtonSize]="IconButtonSize.XL"></sofie-icon-button>
+    </div>
+    <ng-content></ng-content>
+</div>
