--- src/app/producer-shelf/components/producer-shelf/producer-shelf.component.html
@@ -1,17 +1,0 @@
-<div [style.flex-basis.px]="dragOffsetInPixels" [style.height.px]="dragOffsetInPixels">
-    <div #dragHandle class="drag-handle">
-        <sofie-icon-button [iconButton]="IconButton.BARS" [iconButtonSize]="IconButtonSize.XL"></sofie-icon-button>
-    </div>
-    <div
-        *ngFor="let keyBinding of keyBindings" (click)="keyBinding.action()"
-        class="keybinding"
-        style="padding: 12px 20px; margin: 10px;"
-        [class.keybinding--matched]="isKeyBindingMatched(keyBinding)"
-    >
-        {{ keyBinding.label }} ({{ displayKeyBinding(keyBinding) }})
-    </div>
-    <pre style="padding: 20px; color: white;">
-        Pressed:keys
-        {{ pressedKeys }}
-    </pre>
-</div>
