+++ src/app/rundown/components/rundown/rundown.component.html
@@ -15,4 +15,6 @@
     </div>
 </ng-template>
 
-<sofie-producer-shelf></sofie-producer-shelf>
+<sofie-draggable-shelf>
+    <sofie-producer-shelf></sofie-producer-shelf>
+</sofie-draggable-shelf>
